refactor: atualiza importações para utilizar hooks em vez de contextos
